Revert "Merge branch 'master' of https://github.com/aifares/django_prject"
This reverts commit 4e7cf8ea3588aac2018ce071da323ea52437ac64, reversing
changes made to a9b3e2732a3030975c97bd99fac1dfa9fc43f339.

+++ carts/views.py
@@ -169,33 +169,28 @@ def add_to_cart(request, slug):
         new_item = CartItem.objects.create(cart=cart, product=producter)
         single_item = CartItem.objects.filter(cart=cart, product=producter).count()
 
-
+        if qty == 0:
+            remove_from_cart(request, id)
+            print("yooo")
+            zero_qty = False
+            Var_items.delete()
 
         if len(pro_var) > 0 and zero_qty == True:
             for item in pro_var:
-                print("Qty 1")
                 new_item.variation.add(item)
             Var_items = CartItem.objects.filter(cart=cart, product=producter, variation=v)
             Var_items2 = Var_items.filter(variation=p).count()
+
             Check = True
 
 
-        if int(qty) <= 0 and single_item == 1:
-            print("Qty 0")
-            new_item.delete()
-        elif int(qty) <= 0 and (Var_items == 1 or Var_items2 == 1):
-            new_item.delete()
-        else:
-             new_item.quantity = qty
-             new_item.save()
-
-
+        new = []
+        cur = []
         if Var_items2 > 1 and zero_qty == True:
-            print("Qty 2")
             new_item.delete()
             current_item = CartItem.objects.filter(cart=cart, product=producter, variation=v)
             hi = current_item.get(variation=p)
-            if int(qty) <= 0:
+            if qty == '0':
                 hi.delete()
             else:
                 hi.quantity = int(qty)
@@ -203,13 +198,11 @@ def add_to_cart(request, slug):
             Var_items = 0
             Check = True
 
-
-        if not Check:
+        if not Check and zero_qty == True:
             if single_item > 1:
-                print("Qty 3")
                 new_item.delete()
                 current_item = CartItem.objects.get(cart=cart, product=producter)
-                if int(qty) <= 0:
+                if qty == '0':
                     current_item.delete()
                 else:
                     current_item.quantity = int(qty)
@@ -231,4 +224,4 @@ def add_to_cart(request, slug):
     cart.pennies_total = cart.total * 100
     cart.save()
 
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    return redirect(reverse("cart"))
